Rename navbar, media, whiteboard and videodock to the new structure
