+++ bigbluebutton-html5/imports/ui/components/media/component.jsx
