+++ bigbluebutton-html5/imports/ui/components/media/container.jsx
@@ -1,12 +1,12 @@
 import React, { Component, PropTypes } from 'react';
 import { createContainer } from 'meteor/react-meteor-data';
 
-import Media from './Media.jsx';
+import Media from './component.jsx';
 
 import Button from '../button/component';
 
-import WhiteboardContainer from '../whiteboard/WhiteboardContainer';
-import VideoDockContainer from '../video-dock/VideoDockContainer';
+import WhiteboardContainer from '../whiteboard/container';
+import VideoDockContainer from '../video-dock/container';
 
 const defaultProps = {
   overlay: <VideoDockContainer/>,
@@ -34,7 +34,6 @@ class MediaContainer extends Component {
   render() {
     return (
       <Media overlay={this.state.overlay} content={this.state.content}>
-
         <Button
           label="Toggle Layout"
           style={{ position: 'absolute', top: '10px', left: '10px' }}
