+++ bigbluebutton-html5/imports/ui/components/video-dock/component.jsx
