+++ bigbluebutton-html5/imports/ui/components/whiteboard/component.jsx
